requirements.txt
selenium>=4.18.0
pytest>=8.0.0
allure-pytest>=2.13.5
webdriver-manager>=4.0.1
